chore: 디렉토리 구조 수정 (jpa 디렉토리 추가)(#23)
